% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_conservation_calcs.R
\name{add_conservation_calcs}
\alias{add_conservation_calcs}
\title{Add Conservation calculations to temperature and humidity data}
\usage{
add_conservation_calcs(mydata, Temp = "Temp", RH = "RH", ...)
}
\arguments{
\item{mydata}{A dataframe containing temperature and relative humidity data.}

\item{Temp}{Character string name of the temperature column (default "Temp").}

\item{RH}{Character string name of the relative humidity column (default "RH").}

\item{...}{Additional parameters passed to humidity calculation functions.}
}
\value{

}
\description{
Appends columns for conservation-risks: mould risk, preservation indices, equilibrium moisture,
and moisture content for wood to a dataframe with temperature and relative humidity columns.
}
\examples{

# mydata file
filepath <- data_file_path("mydata.xlsx")
mydata <- readxl::read_excel(filepath, sheet = "mydata", n_max = 5)

mydata |> add_conservation_calcs() |> dplyr::glimpse()


}
