% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAH.R
\name{calcAH}
\alias{calcAH}
\title{Calculate absolute humidity}
\usage{
calcAH(Temp, RH)
}
\arguments{
\item{Temp}{Temperature}

\item{RH}{Relative humidity}
}
\value{
AH Absolute humidity
}
\description{
Calculate absolute humidity
}
\examples{
calcAH(20, 50)


}
