% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFP.R
\name{calcFP}
\alias{calcFP}
\title{Calculate Frost Point}
\usage{
calcFP(Temp, RH)
}
\arguments{
\item{Temp}{Temperature (째Celsius)}

\item{RH}{Relative Humidity (0-100\%)}
}
\value{
Tf, Frost Point (째Celsius)
}
\description{
Function to calculate frost point (째C) from temperature (째C) and relative humidity (\%).

This function is under development.
}
\details{
Formula coefficients from Arden Buck equation (1981, 1996) saturation vapor pressure over ice.

\itemize{
  \item a = 6.1115
  \item b = 23.036
  \item c = 279.82
  \item d = 333.7
}
}
\examples{
calcFP(20, 50)
calcFP(0, 50)


# mydata file
filepath <- data_file_path("mydata.xlsx")
mydata <- readxl::read_excel(filepath, sheet = "mydata", n_max = 5)

mydata |> dplyr::mutate(FrostPoint = calcFP(Temp, RH))


}
