% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFtoC.R
\name{calcFtoC}
\alias{calcFtoC}
\title{Convert temperature (F) to temperature (C)}
\usage{
calcFtoC(TempF)
}
\arguments{
\item{TempF}{Temperature (Fahrenheit )}
}
\value{
TempC Temperature (Celsius)
}
\description{
Convert temperature in Fahrenheit to temperature in Celsius
}
\examples{
calcFtoC(32)
calcFtoC(68)

head(mydata) |> dplyr::mutate(TempC = calcFtoC((Temp * 9/5) + 32))


}
