% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIPI.R
\name{calcIPI}
\alias{calcIPI}
\title{Calculate IPI}
\source{
Tim Padfield, 2004
}
\usage{
calcIPI(Temp, RH, EA = 95220)
}
\arguments{
\item{Temp}{Temperature (°Celsius)}

\item{RH}{Relative Humidity (0-100\%)}

\item{EA}{Activation energy E, 95220 J/mol to give the best fit to the cellulose triacetate deterioration data}
}
\value{
IPI formulation of the reaction kinetics, years to a defined state of deterioration
}
\description{
IPI formulation of the reaction kinetics: years to a defined state of deterioration

exp((E - 134.9 × [RH%])/(8.314 × T) + 0.0284 × [RH%] - 28.023)/365
}
\examples{
calcIPI(20, 50)

head(mydata) |> dplyr::mutate(IPI = calcIPI(Temp, RH))


}
