% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLM.R
\name{calcLM}
\alias{calcLM}
\title{Life-time multiplier (Michalski, 2002) for chemical degradation}
\source{
Michalski, 2002
}
\usage{
calcLM(Temp, RH, EA = 100)
}
\arguments{
\item{Temp}{Temperature (Celsius)}

\item{RH}{Relative Humidity (0-100\%)}

\item{EA}{Activation Energy (J/mol) 100 for cellulosic (paper) or 70 yellowing varnish}
}
\value{
Lifetime multiplier
}
\description{
Function to calculate life-time multiplier in J/mol from temperature in Celsius and relative humidity in \% (0-100).

If using a dataframe, columns should ideally be named "Temp" and "RH".
}
\examples{
calcLM(20, 50)

head(mydata) |> dplyr::mutate(LifeTime = calcLM(Temp, RH))


}
\references{
Michalski, S., ‘Double the life for each five-degree drop,
more than double the life for each halving of relative humidity’,
in Preprints of the 13th IcOM-cc Triennial Meeting in rio de Janeiro (22–27 September 2002),
ed. r. Vontobel, James & James, London (2002) Vol. I 66–72.
}
