% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMould_models.R
\name{calcMould_models}
\alias{calcMould_models}
\title{Calculate Mould Growth Probability Using Various Models}
\usage{
calcMould_models(Temp, RH, model = c("Days", "Sedlbauer", "VIR", "Isaksson"))
}
\arguments{
\item{Temp}{Numeric. Temperature (°Celsius)}

\item{RH}{Numeric. Relative Humidity (0-100\%)}

\item{model}{Character. The model to use for calculation.
Options are "Days", "Sedlbauer", "VIR", or "Isaksson".}
}
\value{
Numeric. The result depends on the chosen model:
  \itemize{
    \item Table: returns days to mould
    \item Sedlbauer and VIR: Critical RH for mould growth from temperature
    \item Isaksson: Daily dose (1/t(ms)) / D(rel/day)
  }
}
\description{
This function calculates the probability or critical conditions for mould growth
based on temperature and relative humidity, using one of four different models.
}
\details{
The results are very different so take care in interpretation.
}
\examples{
calcMould_models(20, 80, "Days")
calcMould_models(20, 80, "Sedlbauer")
calcMould_models(25, 70, "VIR")
calcMould_models(22, 85, "Isaksson")


head(mydata) |>
  dplyr::mutate(
    # Days_mould = calcMould_models(Temp, RH, "Days"),
    RH_crit_Sedlbauer = calcMould_models(Temp, RH, "Sedlbauer"),
    RH_crit_VIR = calcMould_models(Temp, RH, "VIR"),
    Mould_daily_dose = calcMould_models(Temp, RH, "Isaksson")
    )



}
