% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMould_VTT.R
\name{calcMould_VTT}
\alias{calcMould_VTT}
\title{Calculate Mould Growth Index (VTT model)}
\usage{
calcMould_VTT(
  Temp,
  RH,
  M_prev = 0,
  sensitivity = "very",
  wood = 0,
  surface = 0
)
}
\arguments{
\item{Temp}{Temperature (°Celsius)}

\item{RH}{Relative Humidity (0-100\%)}

\item{M_prev}{The previous mould index value (default is 0).}

\item{sensitivity}{The sensitivity level of the material to mould growth. Options are 'very', 'sensitive', 'medium', or 'resistant'. Default is 'very'.}

\item{wood}{The wood species; 0 for pine and 1 for spruce. Default is 0.}

\item{surface}{The surface quality; 0 for resawn kiln dried timber and 1 for timber dried under normal kiln drying process. Default is 0.}
}
\value{
Mould growth index
}
\description{
This function calculates the mould growth index on wooden materials based on temperature,
relative humidity, and other factors. It implements the mathematical model
developed by Hukka and Viitanen, which predicts mould growth under varying
environmental conditions.
}
\examples{
calcMould_VTT(Temp = 25, RH = 85)

calcMould_VTT(Temp = 18, RH = 70, M_prev = 2, sensitivity = "medium", wood = 1, surface = 1)

head(mydata) |> dplyr::mutate(MouldIndex = calcMould_VTT(Temp, RH ))



}
\references{
Hukka, A., Viitanen, H. A mathematical model of mould growth on wooden material. Wood Science and Technology 33, 475–485 (1999). https://doi.org/10.1007/s002260050131

Viitanen, Hannu, and Tuomo Ojanen. "Improved model to predict mold growth in building materials." Thermal Performance of the Exterior Envelopes of Whole Buildings X–Proceedings CD (2007): 2-7.
}
