% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMR.R
\name{calcMR}
\alias{calcMR}
\title{Calculate mixing ratio from temperature (C) and relative humidity (\%)}
\usage{
calcMR(Temp, RH, P_tot = 1013.3, B = 621.9907)
}
\arguments{
\item{Temp}{Temperature (Celsius)}

\item{RH}{Relative Humidity (0-100\%)}

\item{P_tot}{Total pressure = 1013.3 (hPa)}

\item{B}{B = 621.9907 g/kg is valid for air}
}
\value{
X Mixing ratio (mass of water vapour / mass of dry gas)
}
\description{
Function to calculate mixing ratio from temperature in Celsius and relative humidity in \% (0-100).

If using a dataframe, columns should ideally be named "Temp" and "RH".

X Mixing ratio (mass of water vapour / mass of dry gas)

Pw = Pws(40C) = 73.75 hPa

X = 621.9907 B7 73.75/(998-73.75) = 49.63 g/kg
}
\examples{
calcMR(20, 50)


}
