% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPws.R
\name{calcPws}
\alias{calcPws}
\title{Calculate water vapour saturation pressure (hPa)}
\usage{
calcPws(Temp)
}
\arguments{
\item{Temp}{Temperature (Celsius)}
}
\value{
Pws, Saturation vapor pressure (hPa). 0.083\% error for T range -20 to +50°C
}
\description{
Function to calculate water vapour saturation pressure in hPa from temperature (°C).
}
\details{
The saturation water vapor pressure (P_ws) is calculated using the following equation:

Where:

\itemize{
  \item $P_{ws}$ is the saturation vapor pressure.
  \item $A$ is a coefficient.
  \item $m$ is a coefficient.
  \item $Tn$ is a constant.
  \item $Temp$ is the temperature in degrees Celsius.
}
}
\note{
If lower accuracy or a limited temperature range can be tolerated a simpler formula
can be used for the water vapour saturation pressure over water (and over ice):

Pws = A \* 10 \^ ( (m \* Temp) / (Temp + Tn) )

August-Roche-Magnus approximation is also included in the source code.
}
\examples{
calcPws(20)

#' # Calculate relative humidity at 50\%RH
calcPw(20, 50) / calcPws(20) * 100

head(mydata) |> dplyr::mutate(Pws = calcPws(Temp))


}
\references{
W. Wagner and A. Pruß:" The IAPWS Formulation 1995 for the Thermodynamic Properties of
Ordinary Water Substance for General and Scientific Use ", Journal of Physical and
Chemical Reference Data, June 2002 ,Volume 31, Issue 2, pp.387535
}
\seealso{
\code{\link{calcMR}} for calculating mixing ratio

\code{\link{calcAD}} for calculating air density

\code{\link{calcPw}} for calculating water vapour pressure

\code{\link{calcPws}} for calculating water vapour saturation pressure
}
