% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRH.R
\name{calcRH}
\alias{calcRH}
\title{Calculate relative humidity from temperature (C) and absolute humidity (g/m^3)}
\usage{
calcRH(Temp = Temp, Abs = Abs)
}
\arguments{
\item{Temp}{Temperature (Celsius)}

\item{Abs}{Absolute Humidity (g/m^3)}
}
\value{
Relative Humidity (0-100\%) with about a 0.1\%RH error
}
\description{
Calculate relative humidity from temperature (C) and absolute humidity (g/m^3)
}
\examples{
calcRH(8.645471, 20)

calcRH(20, calcAH(20, 50))
}
