% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRH_AH.R
\name{calcRH_AH}
\alias{calcRH_AH}
\title{Calculate Relative Humidity from temperature and absolute humidity}
\usage{
calcRH_AH(Temp, Abs)
}
\arguments{
\item{Temp}{Temperature (°Celsius)}

\item{Abs}{Absolute Humidity (g/m³)}
}
\value{
Relative Humidity (0-100\%). Error increases at extreme temperature and relative humidity.
}
\description{
Function to calculate relative humidity (\%) from temperature (°C) and absolute humidity (g/m^3)
}
\examples{
# Calculate RH for temperature of 20°C and absolute humidity of 8.645471 g/m³
calcRH_AH(20, 8.645471)

calcRH_AH(20, calcAH(20, 50))



}
\seealso{
\code{\link{calcTemp}} for calculating temperature

\code{\link{calcRH_DP}} for calculating relative humidity from dew point

\code{\link{calcDP}} for calculating dew point

\code{\link{calcRH}} for calculating relative humidity from absolute humidity
}
