% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRH_DP.R
\name{calcRH_DP}
\alias{calcRH_DP}
\title{Calculate Relative Humidity from temperature and dew point}
\source{
https://bmcnoldy.earth.miami.edu/Humidity.html
}
\usage{
calcRH_DP(Temp, DewP)
}
\arguments{
\item{Temp}{Temperature (Celsius)}

\item{DewP}{Td, Dew Point (Celsius)}
}
\value{
Relative Humidity (0-100\%)
}
\description{
Function to calculate relative humidity (\%) from air temperature (C) and dew point (C)
}
\details{
The relative humidity is calculated using the following equation:

\deqn{RH = 100 \times \frac{\exp(\frac{17.625 \times DewP}{243.04 + DewP})}{\exp(\frac{17.625 \times Temp}{243.04 + Temp})}}

Where:
\itemize{
  \item \eqn{RH} is the relative humidity in percent
  \item \eqn{Temp} is the air temperature in °C
  \item \eqn{DewP} is the dew point temperature in °C
}
}
\examples{
# RH at air tempertaure of 20C and dew point of 15C
calcRH_DP(20, 15)

calcRH_DP(20, calcDew(20, 50))


}
