% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_logfile.R
\name{combine_data}
\alias{combine_data}
\title{Combine TRH data from a list}
\usage{
combine_data(datalist)
}
\arguments{
\item{datalist}{A list of parsed dataframes}
}
\value{
A dataframe containing all timestamped rows in `datalist`
}
\description{
Combines a list of parsed logfiles and returns a
}
\examples{
\donttest{
## Example usage: all_data <- combine_data(datalist)
}

}
