% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConSciR-package.R
\docType{package}
\name{ConSciR-package}
\alias{ConSciR}
\alias{ConSciR-package}
\title{ConSciR: Tools for Conservation Science}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Data science toolbox for conservation science. Tools to help conservators, scientists and engineers in cultural heritage. Data accepted so far - temperature, relative humidity, light. Currently under development.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bhavshah01.github.io/ConSciR/}
  \item Report bugs at \url{https://github.com/BhavShah01/ConSciR/issues}
}

}
\author{
\strong{Maintainer}: Bhavesh Shah \email{bhaveshshah01@gmail.com} (\href{https://orcid.org/0000-0001-8673-0589}{ORCID})

Authors:
\itemize{
  \item Annelies Cosaert (\href{https://orcid.org/0009-0004-1269-4465}{ORCID})
}

Other contributors:
\itemize{
  \item Emily R Long (\href{https://orcid.org/0000-0003-3162-4521}{ORCID}) [contributor]
  \item Marcin Zygmunt (\href{https://orcid.org/0000-0003-1304-8591}{ORCID}) [contributor]
}

}
\keyword{internal}
