% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_file_path.R
\name{data_file_path}
\alias{data_file_path}
\title{Return the file path to data files}
\usage{
data_file_path(path = NULL)
}
\arguments{
\item{path}{Name of file in quotes with extension, e.g. "mydata.xlsx"}
}
\value{
String of the path to the specified file
}
\description{
Access mydata.xlsx and other files in inst/extdata folder
}
\examples{
data_file_path()

data_file_path("mydata.xlsx")

}
