% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_psychrometric.R
\name{graph_psychrometric}
\alias{graph_psychrometric}
\title{Psychrometric chart from temperature and humidity}
\usage{
graph_psychrometric(mydata, Temp, RH, LowT, HighT, LowRH, HighRH)
}
\arguments{
\item{mydata}{A data frame containing temperature and humidity data.}

\item{Temp}{Column name in mydata for temperature values.}

\item{RH}{Column name in mydata for relative humidity values.}

\item{LowT}{Numeric value for low temperature limit.}

\item{HighT}{Numeric value for high temperature limit.}

\item{LowRH}{Numeric value for low relative humidity limit.}

\item{HighRH}{Numeric value for high relative humidity limit.}
}
\value{
A ggplot object representing the psychrometric chart.
}
\description{
Produces a psychrometric chart from temperature and humidity data.
}
\examples{
data <- data.frame(Temp = c(20, 25, 30), RH = c(50, 60, 70))
graph_psychrometric(data, Temp, RH, LowT = 15, HighT = 25, LowRH = 40, HighRH = 60)

graph_psychrometric(head(mydata), Temp, RH, LowT = 15, HighT = 25, LowRH = 40, HighRH = 60)


}
