% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_psychrometric_df.R
\name{graph_psychrometric_df}
\alias{graph_psychrometric_df}
\title{Create a Psychrometric Chart from a dataframe}
\usage{
graph_psychrometric_df(
  mydata,
  Temp = Temp,
  RH = RH,
  y_Axis,
  y_func = calcMR,
  LowT = 16,
  HighT = 25,
  LowRH = 40,
  HighRH = 60,
  Temp_range = c(0, 40),
  y_axis_label = "Y-axis",
  ...
)
}
\arguments{
\item{mydata}{A data frame containing temperature, relative humidity, and y-axis data.}

\item{Temp}{Column name in mydata for temperature values.}

\item{RH}{Column name in mydata for relative humidity values.}

\item{y_Axis}{Column name in mydata for y-axis values.}

\item{y_func}{Function to calculate y-axis values. See above for options and must be
supplied if different from the default: calcMR (mixing ratio).}

\item{LowT}{Numeric value for lower temperature limit of the target range. Default is 16C.}

\item{HighT}{Numeric value for upper temperature limit of the target range. Default is 25C.}

\item{LowRH}{Numeric value for lower relative humidity limit of the target range. Default is 40\%.}

\item{HighRH}{Numeric value for upper relative humidity limit of the target range. Default is 60\%.}

\item{Temp_range}{Numeric vector of length 2 specifying the overall temperature range for the chart.
Default is c(0, 40).}

\item{y_axis_label}{Character string for y-axis label. Default is "Y-axis".}

\item{...}{Additional arguments passed to y_func.}
}
\value{
A ggplot object representing the psychrometric chart.
}
\description{
This function generates a psychrometric chart based on input temperature,
relative humidity, and y-axis data.

Various psychrometric functions can be used for the y-axis if supplied in the data frame:

\itemize{
   \item calcHR: Humidity Ratio (g/kg)
   \item calcMR: Mixing Ratio (g/kg)
   \item calcAH: Absolute Humidity (g/m^3)
   \item calcSH: Specific Humidity (g/kg)
   \item calcAD: Air Density (kg/m^3)
   \item calcDP: Dew Point (C)
   \item calcEnthalpy: Enthalpy (kJ/kg)
   \item calcPws: Saturation vapor pressure (hPa)
   \item calcPw: Water Vapour Pressure (hPa)
   \item calcPI: Preservation Index
   \item calcIPI: Years to Degradation of reference material
   \item calcLM: Lifetime
}
}
\examples{
\dontrun{
# Basic use, don't need to specify y_func as calcMR is default
head(mydata) |>
  mutate(MixingRatio = calcMR(Temp, RH)) |>
  graph_psychrometric_df(y_Axis = MixingRatio)

# Assuming mydata has Temp and RH columns then use a calculation to mutate the y-axis
# Must supply the y_func argument to ensure this function works
head(mydata) |>
  mutate(AbsHum = calcAH(Temp, RH)) |>
  graph_psychrometric_df(y_Axis = AbsHum, y_func = calcAH, y_axis_label = "Absolute Humidity (g/m^3)")
}



}
