% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_TRH.R
\name{graph_TRH}
\alias{graph_TRH}
\title{Function for graphing temperature and humidity data}
\usage{
graph_TRH(data, x_var, y_var, y2_var)
}
\arguments{
\item{mydata}{Dataframe with Date, Temp and RH columns}
}
\value{
Graph of temperature and humidity data
}
\description{
Use this tool to produce a simple temperature and humidity plot
}
\examples{
graph_TRH(mydata)

}
