% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouldtable.R
\docType{data}
\name{mouldtable}
\alias{mouldtable}
\title{Mould datatable}
\format{
A data frame with 1154 rows and 3 columns
\describe{
  \item{TempMould}{Temperature C}
  \item{RHMould}{Relative humidity}
  \item{days_mould}{Days to start of mould growth}
  ...
}
}
\source{
Climate
}
\usage{
mouldtable
}
\description{
A table of days to mould, used for calcMould function.
}
\keyword{datasets}
