% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydata.R
\docType{data}
\name{mydata}
\alias{mydata}
\title{Climate dataset to demonstrate functions}
\format{
A data frame with 35,136 rows and 5 columns:
\describe{
  \item{Site}{Site location name}
  \item{Sensor}{Sensor name, unique to the site}
  \item{Date}{Date is ISOdate time format}
  \item{Temp, RH}{Temperature (C) and relative humidity (\%)}
  ...
}
}
\source{
Climate
}
\usage{
mydata
}
\description{
A climate dataset for use to demonstrate how the functions work.
}
\keyword{datasets}
