% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_logfile.R
\name{parse_brand}
\alias{parse_brand}
\title{Parse datalogger files}
\usage{
parse_brand(directory, Site = "Site", Sensor = "Sensor", brand = FALSE)
}
\arguments{
\item{directory}{A directory of files from a single brand, .csv or .xls only for Rotronic.}

\item{Site}{Character string specifying site name when not recoverable from the file.
Default is "Site".}

\item{Sensor}{Character string specifying sensor name. Default is "Sensor".}

\item{brand}{The logger brand as a string.}
}
\value{
dat, a data frame containing the raw TRH data, with columns
for site, sensor, date, temperature, and relative humidity.
}
\description{
Extracts temperature and humidity data from a directory of logfiles.
}
\examples{
\donttest{
# Example usage:
# parse_brand(dir("logfiles/tinytag", full.names = TRUE), "Anonymous Library", "tinytag")
}

}
