% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ConSciR_app.R
\name{run_ConSciR_app}
\alias{run_ConSciR_app}
\title{Run the ConSciR Shiny Application}
\usage{
run_ConSciR_app()
}
\value{
A Shiny application object that runs interactively.
}
\description{
Launch the ConSciR Shiny app which includes tools for temperature and
relative humidity monitoring such as TRH charts, psychrometric charts,
bivariate plots, mould growth predictions, and a silica gel calculator.

Users can upload CSV or Excel files formatted with "Date", "Temp",
and "RH" columns. The app provides data tidying functions and
downloadable cleaned CSV files.

The silica gel calculator estimates the required amount of silica gel
based on temperature, humidity data, case dimensions, and
air exchange rate (AER) if known.
}
\examples{
if(interactive()) {
    run_ConSciR_app()
}

}
