% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_Mould_app.R
\name{run_Mould_app}
\alias{run_Mould_app}
\title{Run ConSciR Mould Application}
\usage{
run_Mould_app()
}
\value{
Shiny object
}
\description{
Shiny application to calculate mould growth likelihoods.

CSV or Excel formatted data with "Temp" and "RH" columns can be uploaded to the application.
}
\examples{
# run_Mould_app()

}
