% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SilicaGel_app.R
\name{run_SilicaGel_app}
\alias{run_SilicaGel_app}
\title{Run ConSciR Silica Gel Calculator Application}
\usage{
run_SilicaGel_app()
}
\value{
Shiny object
}
\description{
Shiny application to calculate the amount of silica gel required.
Temperature and humidity data from the proposed location, case dimensions,
and air exchange rate (AER) if known, as inputs.

CSV or Excel data with "Date", "Temp" and "RH" columns can be uploaded to the application.
}
\examples{
if(interactive()) {
    run_SilicaGel_app()
}

}
