% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_TidyData_app.R
\name{run_TidyData_app}
\alias{run_TidyData_app}
\title{Run ConSciR Tidying Data Application}
\usage{
run_TidyData_app()
}
\value{
Shiny object
}
\description{
Shiny application to help with tidying data.

CSV or Excel formatted data can be uploaded to the application.
}
\examples{
# run_TidyData_app()

}
