% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_TRHbivariate_app.R
\name{run_TRHbivariate_app}
\alias{run_TRHbivariate_app}
\title{Run ConSciR Temperature and Humidity Application}
\usage{
run_TRHbivariate_app()
}
\value{
Shiny object
}
\description{
Shiny application to plot temperature and humidity data on a bivariate chart.
A summary of the data is also produced by specifying a temperature and humidity box.

CSV or Excel formatted data with "Temp" and "RH" columns can be uploaded to the application.
}
\examples{
# run_Psychrometric_app()

}
