% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_DataUploaderServer.R
\name{shiny_DataUploaderServer}
\alias{shiny_DataUploaderServer}
\title{Server module to read TRH data from an Excel or CSV file into Shiny application}
\usage{
shiny_DataUploaderServer(
  id,
  mydata = NULL,
  Date = "Date",
  Temp_col = "Temp",
  RH_col = "RH"
)
}
\arguments{
\item{id}{The module ID}

\item{mydata}{A data frame containing initial data (if any). Default is NULL.}

\item{Date}{A string specifying the default name of the Date column. Default is "Date".}

\item{Temp_col}{Temp_col A string specifying the default name of the Temperature column. Default is "Temp".}

\item{RH_col}{A string specifying the default name of the Relative Humidity column. Default is "RH".}
}
\value{
A list containing two reactive elements:

  \item{data}{A reactive value containing the uploaded data}
  \item{selected_columns}{A reactive expression that returns a list of the selected column names for Date, Temperature, and RH}
}
\description{
Server module to create a user interface for uploading Excel (.xlsx or .xls) or CSV files
into a Shiny application. It allows users to select specific columns for Date, Temperature,
and Relative Humidity from the uploaded data.
}
\examples{
\dontrun{

# In ui.R
uiOutput("file_upload")
uiOutput("column_Date")
uiOutput("column_Temp")
uiOutput("column_RH")

# In server.R
data_upload <- shiny_DataUploader()

# Access the uploaded data
uploaded_data <- data_upload$data()

# Access the selected column names
selected_cols <- data_upload$selected_columns()
}



}
