% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_DataUploaderUI.R
\name{shiny_DataUploaderUI}
\alias{shiny_DataUploaderUI}
\title{UI module for the data uploader module}
\usage{
shiny_DataUploaderUI(id, ...)
}
\arguments{
\item{id}{A string representing the module ID. This ID is used to namespace
the inputs and outputs of the module, ensuring that they do not conflict
with other modules or UI elements in the application.}

\item{...}{Additional arguements to pass to UI function.}
}
\value{
A tagList containing UI elements for file upload and column selection.
}
\description{
This function creates the user interface components for uploading data files
(CSV or Excel) and selecting relevant columns for Date, Temperature, and
Relative Humidity. It is intended to be used within a Shiny application.
}
\examples{

\dontrun{
# In ui.R or app.R
shiny_DataUploaderUI("dataUpload")
}



}
