% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_DataUploadUI.R
\name{shiny_dataUploadUI}
\alias{shiny_dataUploadUI}
\title{UI Module for Data Upload in Shiny}
\usage{
shiny_dataUploadUI(id)
}
\arguments{
\item{id}{A character string that defines the namespace for the module's UI elements.}
}
\value{
A `tagList` containing a `uiOutput` for file upload. The specific elements
  of this output (such as file input and upload button) are defined in the
  corresponding server function.
}
\description{
This function creates a Shiny UI module for uploading data files. It provides
a file input interface that can be integrated into a larger Shiny application.
}
\examples{
if(interactive()) {

# In a Shiny app:
ui <- fluidPage(
  shiny_dataUploadUI("dataUpload")
)

server <- function(input, output, session) {
  data <- shiny_dataUploadServer("dataUpload")
}

}


}
