% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRHdata.R
\docType{data}
\name{TRHdata}
\alias{TRHdata}
\title{Climate dataset to demonstrate functions}
\format{
A data frame with 35,136 rows and 5 columns:
\describe{
  \item{Site, Sensor}{Sensor location and name}
  \item{Date}{Date is ISOdate time format}
  \item{Temp, RH}{Temperature (C) and relative humidity (\%)}
  ...
}
}
\source{
Climate
}
\usage{
TRHdata
}
\description{
A climate dataset for use to demonstrate how the functions work.
}
\keyword{datasets}
