% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_time_vars.R
\name{add_time_vars}
\alias{add_time_vars}
\title{Add Time Variables to a dataframe}
\usage{
add_time_vars(
  mydata,
  Date = "Date",
  openair_vars = c("seasonyear", "season", "monthyear", "daylight"),
  summer_start = "04-15",
  summer_end = "10-15",
  period_start = NULL,
  period_end = NULL,
  period_label = "Period",
  latitude = 51,
  longitude = -0.5,
  ...
)
}
\arguments{
\item{mydata}{A dataframe containing a date/time column labelled "Date" and "Sensor" column.}

\item{Date}{The name of the date/time column in `mydata` (default "Date").}

\item{openair_vars}{Variables from `openair::cutData()` to add (default includes seasonyear, season, monthyear, daylight).}

\item{summer_start}{Start date for summer season in "MM-DD" format or full date (default "04-15").}

\item{summer_end}{End date for summer season in "MM-DD" format or full date (default "10-15").}

\item{period_start}{Start date of custom period in "MM-DD" format or full date (optional).}

\item{period_end}{End date of custom period in "MM-DD" format or full date (optional).}

\item{period_label}{Label to assign for dates within the custom period, e.g. if there is an Exhibition or a property is open/closed to the public (default "Period").}

\item{latitude}{Latitude for daylight calculations (default 51).}

\item{longitude}{Longitude for daylight calculations (default -0.5).}

\item{...}{Additional arguments passed to `openair::cutData()`.}
}
\value{
Dataframe with appended columns:
  \describe{
    \item{seasonyear, season, monthyear, daylight}{Standard time factors as used in \code{openair::cutData()}.}
    \item{day, hour, dayhour, daymonth, month, year}{Time breakdowns for aggregation and analysis.}
    \item{Season}{Categorises row as "Summer" or "Winter" using user-specified thresholds.}
    \item{Period}{Flags whether each date falls within custom period (e.g. exhibition), with assigned label.}
  }
}
\description{
This function adds multiple time-related variables to a dataframe with a Date column.
It creates standard factors such as season, month-year, day-hour, and determines summer/winter seasonality.
It also allows flexible specification of summer and winter start/end dates, and a custom time period.
}
\examples{

# mydata file
filepath <- data_file_path("mydata.xlsx")
mydata <- readxl::read_excel(filepath, sheet = "mydata", n_max = 5)

mydata |>
  add_time_vars(period_start = "05-01", period_end = "06-30", period_label = "Exhibition") |>
  dplyr::glimpse()

}
