% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAH.R
\name{calcAH}
\alias{calcAH}
\title{Calculate absolute humidity from temperature (C) and relative humidity (\%)}
\usage{
calcAH(Temp, RH)
}
\arguments{
\item{Temp}{Temperature (Celsius)}

\item{RH}{Relative Humidity (0-100\%)}
}
\value{
AH, Absolute Humidity (g/m^3)
}
\description{
Function to calculate absolute humidity (g/m^3) from temperature (C) and relative humidity (\%)
}
\examples{
calcAH(20, 50)


}
