% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAH.R
\name{calcAH}
\alias{calcAH}
\title{Calculate absolute humidity from temperature (C) and relative humidity (\%)}
\usage{
calcAH(Temp = Temp, RH = RH)
}
\arguments{
\item{Temp}{Temperature (Celsius)}

\item{RH}{Relative Humidity (0-100\%)}
}
\value{
Absolute Humidity (g/m^3)
}
\description{
Function to calculate absolute humidity in g per m^3 from temperature in Celsius and relative humidity in \% (0-100).

If using a dataframe, columns should ideally be named "Temp" and "RH".
}
\examples{
calcAH(20, 50)

mydata |> mutate(Abs = calcAH(Temp, RH))
}
