% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDewPoint.R
\name{calcDewPoint}
\alias{calcDewPoint}
\title{Calculate dew point from temperature (C) and relative humidity (\%)}
\usage{
calcDewPoint(Temp = Temp, RH = RH)
}
\arguments{
\item{RH}{Relative humidity (0-100\%)}

\item{TempC}{Temperature (Celsius)}
}
\value{
Dew point (Celsius)
}
\description{
Function to calculate dew point in Celsius from temperature in Celsius and relative humidity in \% (0-100).
Dataframe columns should ideally be named "Temp" and "RH".
}
\examples{
calcDewPoint(20, 50)

}
