% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDP.R
\name{calcDP}
\alias{calcDP}
\title{Calculate dew point}
\source{
https://bmcnoldy.earth.miami.edu/Humidity.html
}
\usage{
calcDP(Temp, RH)
}
\arguments{
\item{Temp}{Temperature (Celsius)}

\item{RH}{Relative Humidity (0-100\%)}
}
\value{
Td, Dew Point (Celsius)
}
\description{
Function to calculate dew point in Celsius from temperature in Celsius and relative humidity in \% (0-100).
}
\details{
Calculation is based on the August-Roche-Magnus approximation, valid for:

\itemize{
   \item 0C < T < 60C
   \item 1% < RH < 100%
   \item 0C < Td < 50C
}

Where:

\itemize{
  \item T_d is the dew point temperature in degrees Celsius.
  \item Temp is the air temperature in degrees Celsius.
  \item RH is the relative humidity in percent 0-100.
}

\eqn{T_d = \frac{243.04 \times \left(\log\left(\frac{RH}{100}\right) + \frac{17.625 \times Temp}{243.04 + Temp}\right)}{17.625 - \log\left(\frac{RH}{100}\right) - \frac{17.625 \times Temp}{243.04 + Temp}}}
}
\note{
The Arden Buck equation is also available in the source R code. Similar results are obtained over -50-100C.
}
\examples{
calcDP(20, 50)

calcDP(20, calcRH(20, calcAH(20, 50)))

head(mydata) |> dplyr::mutate(DewPoint = calcDP(Temp, RH))

}
