% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDP.R
\name{calcDP}
\alias{calcDP}
\title{Calculate dew point from temperature (C) and relative humidity (\%)}
\usage{
calcDP(Temp = Temp, RH = RH)
}
\arguments{
\item{Temp}{number: Temperature (Celsius)}

\item{RH}{number: Relative Humidity (0-100\%)}
}
\value{
number: Dew Point (Celsius)
}
\description{
Function to calculate dew point in Celsius from temperature in Celsius and relative humidity in \% (0-100).

If using a dataframe, columns should ideally be named "Temp" and "RH".

Calculations based on NOAA.
}
\examples{
calc_dew_point(20, 50)

mydata |> mutate(DewPoint = calcDP(Temp, RH))
}
