% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDP.R
\name{calcDP}
\alias{calcDP}
\title{Calculate dew point from temperature (C) and relative humidity (\%)}
\usage{
calcDP(Temp, RH)
}
\arguments{
\item{Temp}{Temperature (Celsius)}

\item{RH}{Relative Humidity (0-100\%)}
}
\value{
Dew Point (Celsius)
}
\description{
Function to calculate dew point in Celsius from temperature in Celsius and relative humidity in \% (0-100).

If using a dataframe, columns should ideally be named "Temp" and "RH".

Calculations based on NOAA.
}
\examples{
calcDP(20, 50)

}
