% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDP.R
\name{calcDP}
\alias{calcDP}
\title{Calculate Dew Point}
\usage{
calcDP(Temp, RH)
}
\arguments{
\item{Temp}{Temperature (°Celsius)}

\item{RH}{Relative Humidity (0-100\%)}
}
\value{
Td, Dew Point (°Celsius)
}
\description{
Function to calculate dew point (°C) from temperature (°C) and relative humidity (\%).
}
\details{
The dew point is calculated using the following equation derived from the August-Roche-Magnus approximation:

\deqn{T_d\left(DewP\right)=\frac{234.04\times\log\left(\frac{RH}{100}\right)+\frac{17.625\times Temp}{243.04+Temp}}{17.625-\log\left(\frac{RH}{100}\right)-\frac{17.625\times Temp}{243.04+Temp}}}

Where:

\itemize{
  \item T_d is the dew point temperature in degrees Celsius.
  \item Temp is the air temperature in degrees Celsius.
  \item RH is the relative humidity in percent.
}

Calculation is valid for:

\itemize{
   \item 0°C < T < 60°C
   \item 1\% < RH < 100\%
   \item 0°C < Td < 50°C
}
}
\note{
The Arden Buck equation is also available in the source R code.
}
\examples{
calcDP(20, 50)

calcDP(20, calcRH_AH(20, calcAH(20, 50)))

head(mydata) |> dplyr::mutate(DewPoint = calcDP(Temp, RH))


}
\references{
Alduchov, O. A., and R. E. Eskridge, 1996: Improved Magnus' form approximation of saturation
vapor pressure. J. Appl. Meteor., 35, 601–609

https://bmcnoldy.earth.miami.edu/Humidity.html
}
\seealso{
\code{\link{calcTemp}} for calculating temperature

\code{\link{calcRH_DP}} for calculating relative humidity from dew point

\code{\link{calcDP}} for calculating dew point

\code{\link{calcRH_AH}} for calculating relative humidity from absolute humidity
}
