% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEnthalpy.R
\name{calcEnthalpy}
\alias{calcEnthalpy}
\title{Calculate enthalpy from temperature (C) and relative humidity (\%)}
\usage{
calcEnthalpy(Temp = Temp, RH = RH)
}
\arguments{
\item{Temp}{Temperature (Celsius)}

\item{RH}{Relative Humidity (0-100\%)}
}
\value{
Enthalpy (kJ/kg)
}
\description{
Function to calculate enthalpy from temperature in Celsius and relative humidity in \% (0-100).

If using a dataframe, columns should ideally be named "Temp" and "RH".
}
\examples{
calcEnthalpy(20, 50)


}
