% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFP.R
\name{calcFP}
\alias{calcFP}
\title{Calculate Frost Point}
\usage{
calcFP(Temp, RH, ...)
}
\arguments{
\item{Temp}{Temperature (째Celsius)}

\item{RH}{Relative Humidity (0-100\%)}

\item{...}{Additional arguments to supply to \code{\link{calcPws}}}
}
\value{
Tf, Frost Point (째Celsius)
}
\description{
Function to calculate frost point (째C) from temperature (째C) and relative humidity (\%).
}
\details{
Formula coefficients

\itemize{
  \item a_ice = 22.452
  \item b_ice = 272.55
}
}
\examples{
calcFP(20, 50)
calcFP(0, 50)

head(mydata) |> dplyr::mutate(FrostPoint = calcFP(Temp, RH))


}
