% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFtoC.R
\name{calcFtoC}
\alias{calcFtoC}
\title{Convert temperature (F) to temperature (C)}
\usage{
calcFtoC(TempF)
}
\arguments{
\item{TempF}{Temperature (Fahrenheit )}
}
\value{
TempC Temperature (Celsius)
}
\description{
Convert temperature in Fahrenheit to temperature in Celsius
}
\examples{
# Fahrenheit from Celsius
calcFtoC(32)
calcFtoC(68)

# mydata file
filepath <- data_file_path("mydata.xlsx")
mydata <- readxl::read_excel(filepath, sheet = "mydata", n_max = 5)

mydata |> dplyr::mutate(TempC = calcFtoC((Temp * 9/5) + 32))


}
