% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMould.R
\name{calcMould}
\alias{calcMould}
\title{Calculate Mould Growth Probability}
\usage{
calcMould(mydata, Date = "Date", Temp = "Temp", RH = "RH")
}
\arguments{
\item{mydata}{A data frame containing temperature and relative humidity data.}

\item{Date}{The name of the column in the data containing date data. Default is "Date".}

\item{Temp}{The name of the column in the data containing temperature data. Default is "Temp".}

\item{RH}{The name of the column in the data containing relative humidity data. Default is "RH".}
}
\value{
A data frame with additional columns:
  \item{time_diff}{Time difference between consecutive measurements}
  \item{mould_prob}{Probability of mould growth for each time step}
  \item{mould}{Cumulative mould growth probability}
}
\description{
This function calculates the probability of mould growth based on temperature and relative humidity data.
}
\examples{
head(mydata) |> calcMould(Date = "Date", Temp = "Temp", RH = "RH")



}
