% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMould_models.R
\name{calcMould_models}
\alias{calcMould_models}
\title{Calculate Mould Growth Probability Using Various Models}
\usage{
calcMould_models(Temp, RH, model = c("Sedlbauer", "VIR", "DP", "Isaksson"))
}
\arguments{
\item{Temp}{Numeric. Temperature in Celsius.}

\item{RH}{Numeric. Relative Humidity as a percentage (0-100).}

\item{model}{Character. The model to use for calculation.
Options are "Sedlbauer", "VIR", "DP", or "Isaksson".}
}
\value{
Numeric. The result depends on the chosen model:
  \itemize{
    \item Sedlbauer and VIR: Critical RH for mould growth
    \item DP: Days to mould growth
    \item Isaksson: Daily dose (1/t(ms)) / D(rel/day)
  }
}
\description{
This function calculates the probability or critical conditions for mould growth
based on temperature and relative humidity, using one of four different models.
}
\details{
The results are very different so take care in interpretation.
}
\examples{
calcMould_models(20, 80, "Sedlbauer")
calcMould_models(25, 70, "VIR")
calcMould_models(30, 75, "DP")
calcMould_models(22, 85, "Isaksson")


}
