% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMR.R
\name{calcMR}
\alias{calcMR}
\title{Calculate mixing ratio}
\usage{
calcMR(Temp, RH, P_atm = 1013.25, B = 621.9907)
}
\arguments{
\item{Temp}{Temperature (Celsius)}

\item{RH}{Relative Humidity (0-100\%)}

\item{P_atm}{Atmospheric pressure = 1013.25 (hPa)}

\item{B}{B = 621.9907 g/kg for air}
}
\value{
X Mixing ratio, mass of water vapour / mass of dry gas (g/kg)
}
\description{
Function to calculate mixing ratio from temperature (°C) and relative humidity (\%).

If using a dataframe, columns should ideally be named "Temp" and "RH".

X Mixing ratio (mass of water vapour / mass of dry gas)

Pw = Pws(40C) = 73.75 hPa

X = 621.9907 B7 73.75/(998-73.75) = 49.63 g/kg
}
\examples{
calcMR(20, 50)

head(mydata) |> dplyr::mutate(MixingRatio = calcMR(Temp, RH))


}
\seealso{
\code{\link{calcMR}} for calculating mixing ratio

\code{\link{calcAD}} for calculating air density

\code{\link{calcPw}} for calculating water vapour pressure

\code{\link{calcPws}} for calculating water vapour saturation pressure
}
