% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPI.R
\name{calcPI}
\alias{calcPI}
\title{Calculate Preservation Index from temperature (C) and relative humidity (\%)}
\source{
Tim Padfield, 2004
}
\usage{
calcPI(Temp, RH)
}
\arguments{
\item{Temp}{Temperature (Celsius)}

\item{RH}{Relative Humidity (0-100\%)}
}
\value{
PI Preservation Index, the expected lifetime 1/k
}
\description{
Chemical degradation of cellulose acetate – the Preservation Index of the
Image Permanence Institute (IPI) (Reilly et al., 1995)

\begin{equation}
 \text{calcPI}(Temp, RH) = \frac{1}/{k}
 \end{equation}

\begin{equation}
 k = RH \cdot 5.9 \times 10^{-12} \cdot \exp\left(\frac{-90300}{8.314 \cdot (Temp + 273.15)}\right)
\end{equation}
}
\examples{
calcPI(20, 50)

head(mydata) |> dplyr::mutate(PI = calcPI(Temp, RH))


}
\references{
Reilly et al., 1995

Padfield, T. 2004. The Preservation Index and The Time Weighted Preservation Index.
https://www.conservationphysics.org/twpi/twpi_01.html
}
