% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPw.R
\name{calcPw}
\alias{calcPw}
\title{Calculate water vapour pressure (hPa)}
\usage{
calcPw(Temp, RH)
}
\arguments{
\item{Temp}{Temperature (Celsius)}

\item{RH}{Relative Humidity (0-100\%)}
}
\value{
Pw, Water Vapour Pressure (hPa)
}
\description{
Function to calculate water vapour pressure in hPa from temperature (°C) and relative humidity (\%).
}
\details{
The water vapor pressure (P_w) is calculated using the following equation:

\deqn{P_w=\frac{P_{ws}\left(Temp\right)\times RH}{100}}

Where:

\itemize{
   \item $P_{ws}$ is the saturation vapor pressure.
   \item $RH$ is the relative humidity in percent.
   \item $Temp$ is the temperature in degrees Celsius.
}
}
\examples{
calcPw(20, 50)

#' #' # Calculate relative humidity at 50\%RH
calcPw(20, 50) / calcPws(20) * 100

head(mydata) |> dplyr::mutate(Pw = calcPws(Temp, RH))


}
\references{
W. Wagner and A. Pru\ß:" The IAPWS Formulation 1995 for the Thermodynamic
Properties of Ordinary Water Substance for General and Scientific Use ",
Journal of Physical and Chemical Reference Data, June 2002 ,Volume 31, Issue 2, pp.387535
}
