% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPw.R
\name{calcPw}
\alias{calcPw}
\title{Calculate water vapour pressure (hPa)}
\usage{
calcPw(Temp, RH)
}
\arguments{
\item{Temp}{Temperature (Celsius)}

\item{RH}{Relative Humidity (0-100\%)}
}
\value{
Pw, Water Vapour Pressure (hPa)
}
\description{
Function to calculate water vapour pressure in hPa between 0C and 373C.
}
\examples{
calcPw(20, 50)


}
\references{
W. Wagner and A. Pru\ß:" The IAPWS Formulation 1995 for the Thermodynamic
Properties of Ordinary Water Substance for General and Scientific Use ",
Journal of Physical and Chemical Reference Data, June 2002 ,Volume 31, Issue 2, pp.387535
}
