% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPws.R
\name{calcPws}
\alias{calcPws}
\title{Calculate water vapour saturation pressure (hPa)}
\usage{
calcPws(Temp)
}
\arguments{
\item{Temp}{Temperature (Celsius)}
}
\value{
Pws, Saturation vapor pressure (hPa). 0.083\% error for T range -20 to +50°C
}
\description{
Function to calculate water vapour saturation pressure in hPa between 0C and 373C.

The saturation vapor pressure (\eqn{P_{ws}}) is calculated using the following equation:

\deqn{P_{ws} = A \times 10^{\left(\frac{m \times Temp}{Temp + Tn}\right)}}

#' Where:
\itemize{
  \item \eqn{P_{ws}} is the saturation vapor pressure.
  \item \eqn{A} is a coefficient.
  \item \eqn{m} is a coefficient.
  \item \eqn{Tn} is a constant.
  \item \eqn{Temp} is the temperature in degrees Celsius.
}
}
\note{
If lower accuracy or a limited temperature range can be tolerated a simpler formula
can be used for the water vapour saturation pressure over water (and over ice):
}
\examples{
calcPws(20)


}
\references{
W. Wagner and A. Pruß:" The IAPWS Formulation 1995 for the Thermodynamic Properties of
Ordinary Water Substance for General and Scientific Use ", Journal of Physical and
Chemical Reference Data, June 2002 ,Volume 31, Issue 2, pp.387535
}
