% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPws.R
\name{calcPws}
\alias{calcPws}
\title{Calculate water vapour pressure (hPa) from temperature (C)}
\usage{
calcPws(Temp = Temp)
}
\arguments{
\item{Temp}{Temperature (Celsius)}
}
\value{
Pws, Saturation vapor pressure (hPa)
}
\description{
Function to calculate water vapour saturation pressure in hPa between 0C and 373C.

If using a dataframe, columns should ideally be named "Temp".
}
\examples{
calcPws(20)
}
\references{
W. Wagner and A. Pruß:" The IAPWS Formulation 1995 for the Thermodynamic Properties of Ordinary Water Substance for General and Scientific Use ", Journal of Physical and Chemical Reference Data, June 2002 ,Volume 31, Issue 2, pp.387535
}
