% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPws.R
\name{calcPws}
\alias{calcPws}
\title{Calculate Water Vapour Saturation Pressure}
\usage{
calcPws(Temp, P_atm = 1013.25, method = "IAPWS")
}
\arguments{
\item{Temp}{Temperature (°Celsius)}

\item{P_atm}{Atmospheric pressure = 1013.25 (hPa)}

\item{method}{Character. Method to use for calculation. Options are "IAPWS" (default) or "Magnus".}
}
\value{
Pws, Saturation vapor pressure (hPa)
}
\description{
Function to calculate water vapour saturation pressure (hPa) from temperature (°C)
using the International Association for the Properties of Water and Steam (IAPWS as default) or
August-Roche-Magnus approximation.

Water vapour saturation pressure is the maximum partial pressure of water vapour that
can be present in gas at a given temperature.
}
\note{
If lower accuracy or a limited temperature range can be tolerated a simpler formula
can be used for the water vapour saturation pressure over water (and over ice):

Pws = 6.116441 x 10^( (7.591386 x Temp) / (Temp + 240.7263) )
}
\examples{
calcPws(20)

calcPws(20, method = "IAPWS")
calcPws(20, method = "Magnus")

# Calculate relative humidity at 50\%RH
calcPw(20, 50) / calcPws(20, method = "IAPWS") * 100
calcPw(20, 50) / calcPws(20, method = "Magnus") * 100

head(mydata) |> dplyr::mutate(Pws = calcPws(Temp))


}
\references{
Wagner, W., & Pru\ß, A. (2002). The IAPWS formulation 1995 for the thermodynamic
properties of ordinary water substance for general and scientific use. Journal of
Physical and Chemical Reference Data, 31(2), 387-535.

Alduchov, O. A., and R. E. Eskridge, 1996: Improved Magnus' form approximation of
saturation vapor pressure. J. Appl. Meteor., 35, 601-609.
}
\seealso{
\code{\link{calcMR}} for calculating mixing ratio

\code{\link{calcAD}} for calculating air density

\code{\link{calcPw}} for calculating water vapour pressure

\code{\link{calcPws}} for calculating water vapour saturation pressure
}
