% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRH.R
\name{calcRH}
\alias{calcRH}
\title{Calculate relative humidity from temperature and absolute humidity}
\usage{
calcRH(Temp, Abs)
}
\arguments{
\item{Temp}{Temperature (Celsius)}

\item{Abs}{Absolute Humidity (g/m^3)}
}
\value{
Relative Humidity (0-100\%). Error increases at extreme temperature and relative humidity.
}
\description{
Function to calculate relative humidity (\%) from temperature (C) and absolute humidity (g/m^3)
}
\details{
The relative humidity is calculated using the following equation:

\deqn{RH = \frac{1}{2165 \times \frac{P_{vs}}{(Temp + 273.15) \times Abs}}}

Where:
\itemize{
  \item \eqn{RH} is the relative humidity (fraction)
  \item \eqn{P_{vs}} is the saturation vapor pressure (Pa)
  \item \eqn{Temp} is the temperature (°C)
  \item \eqn{Abs} is the absolute humidity (g/m^3)
}
}
\examples{
# Calculate RH for temperature of 20°C and absolute humidity of 8.645471 g/m^3
calcRH(20, 8.645471)

calcRH(20, calcAH(20, 50))



}
