% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRH.R
\name{calcRH}
\alias{calcRH}
\title{Calculate relative humidity from temperature (C) and absolute humidity (g/m^3)}
\usage{
calcRH(Temp, Abs)
}
\arguments{
\item{Temp}{Temperature (Celsius)}

\item{Abs}{Absolute Humidity (g/m^3)}
}
\value{
Relative Humidity (0-100\%). At extreme high and low humidity the error increases.
}
\description{
Function to calculate relative humidity (\%) from temperature (C) and absolute humidity (g/m^3)
}
\examples{
calcRH(20, 8.645471)

calcRH(20, calcAH(20, 50))


}
