% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRH_DP.R
\name{calcRH_DP}
\alias{calcRH_DP}
\title{Calculate Relative Humidity from temperature and dew point}
\usage{
calcRH_DP(Temp, DewP)
}
\arguments{
\item{Temp}{Temperature (°Celsius)}

\item{DewP}{Td, Dew Point (°Celsius)}
}
\value{
Relative Humidity (0-100\%)
}
\description{
Function to calculate relative humidity (\%) from temperature (°C) and dew point (°C)
}
\details{
The relative humidity is calculated using the following equation derived from the August-Roche-Magnus approximation:

\deqn{RH=100\times\frac{\exp\left(\frac{\left(17.625\times DewP\right)}{\left(243.04+DewP\right)}\right)}{\exp\left(\frac{\left(17.625\times Temp\right)}{\left(243.04+Temp\right)}\right)}}

Where:

\itemize{
  \item RH is the relative humidity in percent
  \item Temp is the air temperature in °C
  \item DewP is the dew point temperature in °C
}
}
\examples{
# RH at air tempertaure of 20C and dew point of 15°C
calcRH_DP(20, 15)

calcRH_DP(20, calcDP(20, 50))

head(mydata) |> dplyr::mutate(DewPoint = calcDP(Temp, RH), RH2 = calcRH_DP(Temp, DewPoint))



}
\references{
Alduchov, O. A., and R. E. Eskridge, 1996: Improved Magnus' form approximation of
saturation vapor pressure. J. Appl. Meteor., 35, 601-609.

https://bmcnoldy.earth.miami.edu/Humidity.html
}
\seealso{
\code{\link{calcTemp}} for calculating temperature

\code{\link{calcDP}} for calculating dew point

\code{\link{calcRH_AH}} for calculating relative humidity from absolute humidity

\code{\link{calcRH_DP}} for calculating relative humidity from dew point
}
