% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRH_DP.R
\name{calcRH_DP}
\alias{calcRH_DP}
\title{Calculate Relative Humidity from temperature and dew point}
\usage{
calcRH_DP(Temp, DewP, method = c("Magnus", "Buck"))
}
\arguments{
\item{Temp}{Temperature (°Celsius)}

\item{DewP}{Td (DP), Dew Point (°Celsius)}

\item{method}{Calculation method: either \code{"Magnus"} or \code{"Buck"}.
Defaults to \code{"Magnus"}.}
}
\value{
Relative Humidity (0-100\%)
}
\description{
Function to calculate relative humidity (\%) from temperature (°C) and dew point (°C)
}
\details{
This function supports two methods for relative humidity calculation:
\itemize{
  \item \code{"Magnus"} (default): Uses the August-Roche-Magnus approximation,
  valid for 0°C < Temp < 60°C and 1\% < RH < 100\%.
  \item \code{"Buck"}: Uses the Arden Buck equation with Bögel modification,
  valid for -30°C < Temp < 60°C and 1\% < RH < 100\%.
}
The methods calculate temperature based on vapor pressure and saturation vapour
pressure relationships.
The Magnus method is chosen as the default because it is more stable
when used with the \code{\link{calcDP}} and \code{\link{calcTemp}} functions.
}
\examples{
# RH at air tempertaure of 20°C and dew point of 15°C
calcRH_DP(20, 15)
calcRH_DP(20, 15, method = "Buck")

calcRH_DP(20, calcDP(20, 50))

calcRH_DP(20, calcDP(20, 50, method = "Buck"), method = "Buck")

head(mydata) |>
  dplyr::mutate(
    DewPoint = calcDP(Temp, RH),
    RH_default = calcRH_DP(Temp, DewPoint),
    RH_Buck = calcRH_DP(Temp, DewPoint, method = "Buck"))



}
\references{
Alduchov, O. A., and R. E. Eskridge, 1996: Improved Magnus' form approximation of saturation
vapor pressure. J. Appl. Meteor., 35, 601–609

Buck, A. L., 1981: New Equations for Computing Vapor Pressure and Enhancement Factor.
J. Appl. Meteor. Climatol., 20, 1527–1532,
https://doi.org/10.1175/1520-0450(1981)020<1527:NEFCVP>2.0.CO;2.

Buck (1996), Buck (1996), Buck Research CR-1A User's Manual, Appendix 1.

https://bmcnoldy.earth.miami.edu/Humidity.html
}
\seealso{
\code{\link{calcTemp}} for calculating temperature

\code{\link{calcDP}} for calculating dew point

\code{\link{calcRH_AH}} for calculating relative humidity from absolute humidity

\code{\link{calcRH_DP}} for calculating relative humidity from dew point
}
