% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSH.R
\name{calcSH}
\alias{calcSH}
\title{Calculate Specific Humidity}
\usage{
calcSH(Temp, RH, P_atm = 1013.25, ...)
}
\arguments{
\item{Temp}{Temperature (°Celsius)}

\item{RH}{Relative Humidity (0-100\%)}

\item{P_atm}{Atmospheric pressure = 1013.25 (hPa)}

\item{...}{Additional arguments to supply to \code{\link{calcPws}}}
}
\value{
SH Specific Humidity (g/kg)
}
\description{
Function to calculate the specific humidity (g/kg) from temperature (°C) and relative humidity (\%).

Specific humidity is the ratio of the mass of water vapor to the mass of air.
}
\examples{
calcSH(20, 50)

head(mydata) |> dplyr::mutate(SpecificHumidity = calcSH(Temp, RH))



}
\references{
Wallace, J.M. and Hobbs, P.V. (2006). Atmospheric Science: An Introductory Survey.
Academic Press, 2nd edition.
}
\seealso{
\code{\link{calcAD}} for calculating air density

\code{\link{calcAH}} for calculating absolute humidity

\code{\link{calcPw}} for calculating water vapour pressure

\code{\link{calcPws}} for calculating water vapour saturation pressure
}
