% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTemp.R
\name{calcTemp}
\alias{calcTemp}
\title{Calculate Temperature from relative humidity and dew point}
\usage{
calcTemp(RH, DewP)
}
\arguments{
\item{RH}{Relative Humidity (0-100\%)}

\item{DewP}{Td, Dew Point (°Celsius)}
}
\value{
Temp, Temperature (°Celsius)
}
\description{
This function calculates the temperature (°C) from relative humidity (\%) and dew point temperature (°C).
}
\details{
The temperature is calculated using the following equation derived from the August-Roche-Magnus approximation:

\deqn{TempC=\frac{243.04\times\left(\frac{17.625\times DewP}{243.04+DewP}-\log\left(\frac{RH}{100}\right)\right)}{17.625+\log\left(\frac{RH}{100}\right)-\frac{17.625\times DewP}{243.04+DewP}}}

Where:

\itemize{
  \item Temp is the calculated temperature in Celsius
  \item RH is the relative humidity in percent
  \item DewP is the dew point temperature in Celsius
}
}
\examples{
# Calculate temperature for RH of 50\% and dew point of 15°C
calcTemp(50, 15)

calcTemp(50, calcDP(20, 50))

head(mydata) |> dplyr::mutate(DewPoint = calcDP(Temp, RH), Temp2 = calcTemp(RH, DewPoint))


}
\seealso{
\code{\link{calcTemp}} for calculating temperature

\code{\link{calcDP}} for calculating dew point

\code{\link{calcRH_DP}} for calculating relative humidity from dew point

\code{\link{calcRH_AH}} for calculating relative humidity from absolute humidity
}
