% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ConSciR_app.R
\name{run_ConSciR_app}
\alias{run_ConSciR_app}
\title{Run ConSciR Shiny Application}
\usage{
run_ConSciR_app()
}
\value{
Shiny object
}
\description{
Run ConSciR Shiny Application
}
\examples{
# run_ConSciR_app()

}
