% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_Mould_app.R
\name{run_Mould_app}
\alias{run_Mould_app}
\title{Run ConSciR Mould Application}
\usage{
run_Mould_app()
}
\value{
Shiny object
}
\description{
Shiny application to upload data to estimate mould growth predictions.

CSV or Excel formatted data with "Temp" and "RH" columns can be uploaded to the application.

Functions available:

\itemize{
 \item calcMould_Zeng: Mould Growth Rate Limits
 \item calcMould_VTT: Mould Growth Index (VTT model)
 }
}
\examples{
if(interactive()) {
    run_Mould_app()
}


}
