% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_Psychrometric_app.R
\name{run_Psychrometric_app}
\alias{run_Psychrometric_app}
\title{Run ConSciR Psychrometric Application}
\usage{
run_Psychrometric_app()
}
\value{
Shiny object
}
\description{
Run ConSciR Psychrometric Application
}
\examples{
# run_Psychrometric_app()

}
