% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_DataUploadUI.R
\name{shiny_dataUploadUI}
\alias{shiny_dataUploadUI}
\title{UI Module for Data Upload in Shiny with Averaging Option}
\usage{
shiny_dataUploadUI(id)
}
\arguments{
\item{id}{A character string that defines the namespace for the module's UI elements.}
}
\value{
A `tagList` containing UI elements for file upload and averaging option.
}
\description{
This function creates a Shiny UI module for uploading data files and optionally
selecting hourly averaging. It provides a file input interface and averaging
selection that can be integrated into a larger Shiny application.
}
\examples{
if(interactive()) {

# In a Shiny app:
ui <- fluidPage(
  shiny_dataUploadUI("dataUpload")
)

server <- function(input, output, session) {
  data <- shiny_dataUploadServer("dataUpload")
}

}

}
